fb: restore navbar compatibility with several xposed modules I am tired of the complains....
Xposed use the emulator overwrite to enable/disable navbar...this results
into showing up navbar on reboot even if the user has set it to off in settings.

This fixes the issue and still keeps for emulator tests the basic overwrite functionallity
alive as long the provider value was not set.

We may should reflect that in settings as well.

Change-Id: I99f04da8110bbcfcdba62604610d1f4330aea4d0

Conflicts:
	policy/src/com/android/internal/policy/impl/PhoneWindowManager.java